add test case for null value,rename function,spacify which exception to throw
